drivers: entropy: cc310: Workaround for spm secure services limitation
`spm_request_random_number()` requires that the application requests
exactly 144 random bytes, no different. As it's called directly in the
cc310 entropy driver, this is an unacceptable limitation, as there's
at the `entropy_get_entropy()` function, which is used in various
different libraries.

This commit works around this limitation in the
`entropy_cc310_rng_get_entropy` implementation.

Signed-off-by: Robert Lubos <robert.lubos@nordicsemi.no>